# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BitaSoft Api Documentation",
        "description": "BitaSoft Api Documentation",
        "contact": {
            "name": "Bita Soft",
            "email": "bitasoft22@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/product/list": {
            "get": {
                "tags": [
                    "List"
                ],
                "summary": "Get list of Lists",
                "description": "Get list of Lists",
                "operationId": "1525185e10ba4b2b521535374d8de3ee",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "the page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "order accepts 'asc' or 'desc'",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "List"
                ],
                "summary": "Store Lists in DB",
                "description": "Store Lists in DB",
                "operationId": "86c87195176e7c7c02a39b896440129c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "safeEmail",
                                    "password",
                                    "tollFreePhoneNumber",
                                    "jobTitle",
                                    "city",
                                    "country",
                                    "latitude",
                                    "longitude",
                                    "timeZone"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "safeEmail": {
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "tollFreePhoneNumber": {
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "jobTitle": {
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "city": {
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "country": {
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "latitude": {
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "longitude": {
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "timeZone": {
                                        "type": "string",
                                        "format": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/product/list/{id}": {
            "get": {
                "tags": [
                    "List"
                ],
                "summary": "Get Lists Detail",
                "description": "Get Lists Detail",
                "operationId": "a4279a7dc5993eed106e0ff1c8647b60",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of Lists",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "List"
                ],
                "summary": "Update Lists in DB",
                "description": "Update Lists in DB",
                "operationId": "cb134208a40d92ff9af5debbd9aea4b1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of Lists",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "safeEmail",
                                    "password",
                                    "tollFreePhoneNumber",
                                    "jobTitle",
                                    "city",
                                    "country",
                                    "latitude",
                                    "longitude",
                                    "timeZone"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "safeEmail": {
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "tollFreePhoneNumber": {
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "jobTitle": {
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "city": {
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "country": {
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "latitude": {
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "longitude": {
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "timeZone": {
                                        "type": "string",
                                        "format": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "List"
                ],
                "summary": "Delete Lists",
                "description": "Delete Lists",
                "operationId": "aec7b3d28398665f5cb719b037c9b0a9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of Lists",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}